address comment regarding test
