+++ src/api.v2/controllers/__mocks__/accessController.js
@@ -1,28 +1,4 @@
-const { NotFoundError } = require('../../../utils/responses');
-
-const mockUsersList = [
-  {
-    name: 'Mock Admin',
-    email: 'admin@example.com',
-    role: 'admin',
-  },
-  {
-    name: 'Mock User',
-    email: 'user@example.com',
-    role: 'owner',
-  },
-];
-
-const mockGetExperimentUsers = jest.fn((req, res) => {
-  const { experimentId } = req.params;
-
-  if (experimentId === 'nonExistentExperimentId') {
-    throw new NotFoundError('Experiment not found');
-  }
-
-  res.json(mockUsersList);
-  return Promise.resolve();
-});
+const mockGetExperimentUsers = jest.fn();
 
 module.exports = {
   getExperimentUsers: mockGetExperimentUsers,
