+++ tests/api.v2/routes/access.test.js
@@ -3,8 +3,24 @@ const express = require('express');
 const request = require('supertest');
 const expressLoader = require('../../../src/loaders/express');
 
+const accessController = require('../../../src/api.v2/controllers/accessController');
+const { NotFoundError } = require('../../../src/utils/responses');
+
 jest.mock('../../../src/api.v2/middlewares/authMiddlewares');
 jest.mock('../../../src/api.v2/controllers/accessController');
+
+const mockUsersList = [
+  {
+    name: 'Mock Admin',
+    email: 'admin@example.com',
+    role: 'admin',
+  },
+  {
+    name: 'Mock User',
+    email: 'user@example.com',
+    role: 'owner',
+  },
+];
 
 describe('User access endpoint', () => {
   let app;
@@ -20,6 +36,11 @@ describe('User access endpoint', () => {
   });
 
   it('Getting list of users to an existing experiment returns 200', async (done) => {
+    accessController.getExperimentUsers.mockImplementationOnce((req, res) => {
+      res.json(mockUsersList);
+      Promise.resolve();
+    });
+
     request(app)
       .get('/v2/access/mockExperimentId')
       .expect(200)
@@ -32,6 +53,10 @@ describe('User access endpoint', () => {
   });
 
   it('Getting list of users to an unexisting experiment returns 404', async (done) => {
+    accessController.getExperimentUsers.mockImplementationOnce(() => {
+      throw new NotFoundError('Experiment not found');
+    });
+
     request(app)
       .get('/v2/access/nonExistentExperimentId')
       .expect(404)
